BACKLOG-10461: move MediaPicker and ContentPicker stuff in a dedicated folder
