+++ src/javascript/EditPanelContainer/EditPanel/EditPanelContent/FormBuilder/SelectorTypes/Picker/ContentPicker/ContentPickerFilled/ContentPickerFilled.jsx
@@ -2,9 +2,9 @@ import React, {useState} from 'react';
 import {ProgressOverlay} from '@jahia/react-material';
 import * as PropTypes from 'prop-types';
 import {translate} from 'react-i18next';
-import {Picker} from '../../../../../../../DesignSystem/Picker';
+import {Picker} from '../../../../../../../../DesignSystem/Picker';
 import {ContentPickerFilledQuery} from './ContentPickerFilled.gql-queries';
-import {encodeJCRPath} from '../../../../../EditPanel.utils';
+import {encodeJCRPath} from '../../../../../../EditPanel.utils';
 import {useQuery} from 'react-apollo-hooks';
 import {DisplayActions} from '@jahia/react-material';
 import IconButton from '@material-ui/core/IconButton';
