+++ src/javascript/EditPanelContainer/EditPanel/EditPanelContent/FormBuilder/SelectorTypes/Picker/ContentPicker/ContentTable/ContentTable.container.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import DSContentTable from '../../../../../../../DesignSystem/ContentTable/ContentTable';
+import DSContentTable from '../../../../../../../../DesignSystem/ContentTable/ContentTable';
 import {useQuery} from 'react-apollo-hooks';
 import {translate} from 'react-i18next';
 import {ProgressOverlay} from '@jahia/react-material';
