+++ src/javascript/EditPanelContainer/EditPanel/EditPanelContent/FormBuilder/SelectorTypes/Picker/MediaPicker/ImageListQuery/ImageListQuery.jsx
@@ -3,8 +3,8 @@ import {ProgressOverlay} from '@jahia/react-material';
 import * as PropTypes from 'prop-types';
 import {useQuery} from 'react-apollo-hooks';
 import {translate} from 'react-i18next';
-import {ImageList} from '../../../../../../../DesignSystem/ImageList';
-import {encodeJCRPath} from '../../../../../EditPanel.utils';
+import {ImageList} from '../../../../../../../../DesignSystem/ImageList';
+import {encodeJCRPath} from '../../../../../../EditPanel.utils';
 import {MediaPickerImages} from './ImageListQuery.gql-queries';
 
 const ImageListQueryCmp = ({
