+++ src/javascript/EditPanelContainer/EditPanel/EditPanelContent/FormBuilder/SelectorTypes/Picker/MediaPicker/MediaPicker.spec.js
