+++ src/javascript/EditPanelContainer/EditPanel/EditPanelContent/FormBuilder/SelectorTypes/Picker/MediaPicker/MediaPickerEmpty/MediaPickerEmpty.spec.jsx
