+++ src/javascript/EditPanelContainer/EditPanel/EditPanelContent/FormBuilder/SelectorTypes/Picker/MediaPicker/MediaPickerFilled/MediaPickerFilled.jsx
@@ -2,8 +2,8 @@ import React, {useState} from 'react';
 import {ProgressOverlay} from '@jahia/react-material';
 import * as PropTypes from 'prop-types';
 import {translate} from 'react-i18next';
-import {Picker} from '../../../../../../../DesignSystem/Picker';
-import {encodeJCRPath} from '../../../../../EditPanel.utils';
+import {Picker} from '../../../../../../../../DesignSystem/Picker';
+import {encodeJCRPath} from '../../../../../../EditPanel.utils';
 import {MediaPickerFilledQuery} from './MediaPickerFilled.gql-queries';
 import {MediaPickerDialog} from '../MediaPickerDialog';
 import {DisplayActions} from '@jahia/react-material';
