+++ src/javascript/EditPanelContainer/EditPanel/EditPanelContent/FormBuilder/SelectorTypes/Picker/index.js
@@ -1,0 +1,3 @@
+import pickerConfigs from './pickerConfigs';
+
+export default pickerConfigs;
