+++ src/javascript/EditPanelContainer/EditPanel/EditPanelContent/FormBuilder/SelectorTypes/Picker/pickerConfigs.spec.js
@@ -1,21 +1,21 @@
-import pickersConfig from './pickersConfig';
+import pickerConfigs from './pickerConfigs';
 
 describe('Pickers Configs', () => {
     describe('resolveComponent', () => {
         it('should always return a component', () => {
-            const cmpDefinition = pickersConfig.resolveComponent();
+            const cmpDefinition = pickerConfigs.resolveComponent();
             expect(cmpDefinition.cmp).toBeDefined();
             expect(cmpDefinition.key).toBeDefined();
         });
     });
     describe('resolveConfig', () => {
         it('should always return a config', () => {
-            const config = pickersConfig.resolveConfig();
+            const config = pickerConfigs.resolveConfig();
             expect(config).toBeDefined();
         });
         it('should override correctly', () => {
             const formDefinition = {valueConstraints: [{value: {string: 'jmix:droppableContent'}}]};
-            const config = pickersConfig.resolveConfig([{name: 'type', value: 'editorial'}], formDefinition);
+            const config = pickerConfigs.resolveConfig([{name: 'type', value: 'editorial'}], formDefinition);
             expect(config.selectableTypesTable).toEqual(['jmix:droppableContent']);
         });
     });
